Fix : Relocate Files/Folder
