+++ src/App.tsx
@@ -3,18 +3,18 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import Layout from "./components/layout/Layout";
 import Home from "./routes/Home";
-import Notice from "./routes/Notice";
-import ActivityRecommed from "./routes/ActivityRecommed";
-import OutfitByTemp from "./routes/OutfitByTemp";
+import Notice from "./routes/info/Notice";
+import ActivityRecommed from "./routes/recommend/ActivityRecommed";
+import OutfitByTemp from "./routes/recommend/OutfitByTemp";
 import Favorites from "./routes/Favorites";
 import LocalPride from "./routes/LocalPride";
 import SensitiveCheck from "./routes/SensitiveCheck";
-import Setting from "./routes/Setting";
-import Contact from "./routes/Contact";
-import OutfitDaily from "./routes/OutfitDaily";
+import Setting from "./routes/info/Setting";
+import Contact from "./routes/info/Contact";
+import OutfitDaily from "./routes/recommend/OutfitDaily";
 import Collection from "./routes/Collection";
 import DaliyMission from "./routes/DailyMission";
-import GameList from "./routes/GameList";
+import GameList from "./routes/games/GameList";
 import WeatherInfo from "./routes/WeatherInfo";
 const App: React.FC = () => {
   return (
