+++ src/routes/info/Notice.tsx
