+++ src/routes/info/styles/Notice.css
